split up commands/create
