+++ src/lib/commands.ts
@@ -1,11 +1,12 @@
 import { ExtensionContext } from 'vscode';
+import { registerCommands as registerAccountCommands } from './commands/account';
 import { registerCommands as registerAuthCommands } from './commands/auth';
 import { registerCommands as registerConfigCommands } from './commands/config';
-import { registerCommands as registerNotificationCommands } from './commands/notifications';
-import { registerCommands as registerTerminalCommands } from './commands/terminal';
-import { registerCommands as registerCreateCommands } from './commands/create';
-import { registerCommands as registerAccountCommands } from './commands/account';
 import { registerCommands as registerGlobalStateCommands } from './commands/globalState';
+import { registerCommands as registerNotificationCommands } from './commands/notifications';
+import { registerCommands as registerServerlessCommands } from './commands/serverless';
+import { registerCommands as registerTemplateCommands } from './commands/templates';
+import { registerCommands as registerTerminalCommands } from './commands/terminal';
 
 export const registerCommands = (
   context: ExtensionContext,
@@ -14,8 +15,9 @@ export const registerCommands = (
   registerAccountCommands(context);
   registerAuthCommands(context, rootPath);
   registerConfigCommands(context);
-  registerNotificationCommands(context);
-  registerCreateCommands(context);
-  registerTerminalCommands(context);
   registerGlobalStateCommands(context);
+  registerNotificationCommands(context);
+  registerServerlessCommands(context);
+  registerTemplateCommands(context);
+  registerTerminalCommands(context);
 };
